[Gold IV] Title: 기차 여행, Time: 424 ms, Memory: 60888 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/10713. 기차 여행/기차 여행.java
@@ -48,7 +48,7 @@ public class Main {
     	
     	for (int i = 1; i < N; i++) {
 			cnt[i]+=cnt[i-1];
-			ans+=Math.min(ticket[i]*cnt[i], ic[i]+cost[i]*cnt[i]);
+			ans+=Math.min((long) ticket[i]*cnt[i], (long) ic[i]+(long) cost[i]*cnt[i]);
 		}
     }
